feat: v1 AWS 배포
